Working primarily on getting unit tests working again. May have resolved issue with Cache and Cached Lists.
