+++ pom.xml
@@ -4,7 +4,7 @@
 	<groupId>nz.net.gravity</groupId>
 	<artifactId>gravity</artifactId>
 	<packaging>war</packaging>
-	<version>3.0.0</version>
+	<version>3.0.1-SNAPSHOT</version>
 	<name>Gravity Workflow Automation</name>
 
 	<properties>
