+++ src/main/java/nz/net/orcon/kanban/tools/MemoryRepositoryFactory.java
@@ -1,0 +1,14 @@
+package nz.net.orcon.kanban.tools;
+
+import javax.jcr.Repository;
+import org.apache.jackrabbit.oak.Oak;
+import org.apache.jackrabbit.oak.jcr.Jcr;
+
+public class MemoryRepositoryFactory implements RepositoryFactory {
+	
+	@Override
+	public Repository getFactoryInstance() {
+		return new Jcr(new Oak()).createRepository();
+	}
+
+}
