+++ src/main/java/nz/net/orcon/kanban/tools/RepositoryFactory.java
@@ -1,0 +1,9 @@
+package nz.net.orcon.kanban.tools;
+
+import javax.jcr.Repository;
+
+public interface RepositoryFactory {
+	
+	public Repository getFactoryInstance() throws Exception;
+
+}
