+++ src/main/webapp/WEB-INF/spring/app/repository.xml
@@ -15,11 +15,15 @@
 		<constructor-arg value="${repository.url}"/>
 	</bean>
 	 -->
+	 
+	<bean id="repositoryFactory" class="nz.net.orcon.kanban.tools.MongoRepositoryFactory">
+		<property name="host" value="${repository.url}"/>
+	</bean>
 	 	
 	<bean id="ocmFactory" class="nz.net.orcon.kanban.tools.OcmMapperFactory">
+		<property name="repositoryFactory" ref="repositoryFactory"/>
 		<property name="user" value="admin"/>
 		<property name="password" value="admin"/>
-		<property name="host" value="${repository.url}"/>
 		<property name="domainPackage" value="nz.net.orcon.kanban.model"/>
 		<property name="mainNodes">
 			<list>
