+++ src/test/resources/test-cache.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:mvc="http://www.springframework.org/schema/mvc"
+	xmlns:context="http://www.springframework.org/schema/context" xmlns:tx="http://www.springframework.org/schema/tx"
+	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
+		http://www.springframework.org/schema/mvc http://www.springframework.org/schema/mvc/spring-mvc-3.0.xsd	
+		http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.0.xsd
+		http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd"
+	default-autowire="byName">
+	
+	<bean id="gravityConfiguration"
+		class="org.springframework.beans.factory.config.PropertiesFactoryBean">
+		<property name="location" value="file:///etc/gravity.conf" />
+	</bean>
+
+	<context:property-placeholder properties-ref="gravityConfiguration" />
+	
+	<bean id="awsCredential"  class="com.amazonaws.auth.BasicAWSCredentials">
+		<constructor-arg value="${aws.keyid}"/>
+		<constructor-arg value="${aws.secret}"/>
+	</bean>
+	
+	<bean id="awsLoadBalanceClient" class="com.amazonaws.services.elasticloadbalancing.AmazonElasticLoadBalancingClient">
+		<constructor-arg ref="awsCredential"/>
+		<property name="endpoint" value="https://elasticloadbalancing.ap-southeast-2.amazonaws.com"></property>
+	</bean>
+
+	<bean id="awsIdentityClient" class="com.amazonaws.services.identitymanagement.AmazonIdentityManagementClient">
+		<constructor-arg ref="awsCredential"/>
+	</bean>
+	
+	<bean id="awsCloudFrontClient" class="com.amazonaws.services.cloudfront.AmazonCloudFrontClient">
+		<constructor-arg ref="awsCredential"/>
+		<property name="endpoint" value="https://cloudfront.amazonaws.com"></property>
+	</bean>
+	
+	<bean id="awsAutomationAction" class="nz.net.orcon.kanban.automation.actions.AwsAutomationAction"/>	
+	
+</beans>
