+++ src/test/resources/test-controllers.xml
@@ -19,6 +19,7 @@
 	
 	<context:property-placeholder properties-ref="kanbanConfiguration"/>
 	
+	<import resource="test-event-config.xml" />
 	<import resource="test-repository.xml" />
 				
 	<context:component-scan base-package="nz.net.orcon.kanban.controllers" />
@@ -30,6 +31,14 @@
 	<bean id="validator" class="org.springframework.validation.beanvalidation.LocalValidatorFactoryBean" />
 		
 	<bean id="messageFactory" class="org.springframework.ws.soap.saaj.SaajSoapMessageFactory" />
+	
+	<bean id="securityTool" class="nz.net.orcon.kanban.security.MockSecurityTool"/>
+	
+	<bean id="clusterManager" class="nz.net.orcon.kanban.automation.MockClusterManager"/>
+	
+	<bean id="jcrObserver" class="nz.net.orcon.kanban.automation.JcrObserver"/>
+	
+	<bean id="testBoardTool" class="nz.net.orcon.kanban.controllers.TestBoardTool"/>
 	
 	<bean id="listTools" class="nz.net.orcon.kanban.tools.ListTools" >
 		<property name="dateInterpreter" ref="dateInterpreter" />
@@ -45,9 +54,11 @@
 		</property>
 	</bean> 
 	
-	<bean id="cacheManager" class="nz.net.orcon.kanban.automation.MockCacheInvalidationManager"/>
+	<bean id="cacheManager" class="nz.net.orcon.kanban.automation.MockCacheManager"/>
 	
 	<bean id="cardsLockCache" class="nz.net.orcon.kanban.controllers.MockCardLockCache"/>
+	
+	<bean id="cardTools" class="nz.net.orcon.kanban.tools.CardToolsImpl" />
 	
 	<bean id="xstreamMarshaller" class="org.springframework.oxm.xstream.XStreamMarshaller">
         <property name="aliases">
